Move plotting function to graph
