+++ pyart/graph/__init__.py
@@ -60,5 +60,6 @@ from .gridmapdisplay import GridMapDisplay
 from .gridmapdisplay_basemap import GridMapDisplayBasemap
 from .radarmapdisplay import RadarMapDisplay
 from .radarmapdisplay_basemap import RadarMapDisplayBasemap
+from .convstrat_scheme_plot import plot_convstrat_scheme
 
 __all__ = [s for s in dir() if not s.startswith('_')]
